работа по замечаниям
