+++ ImageFeed/Addons/AccessibilityIdentifiers.swift
@@ -1,0 +1,24 @@
+//
+//  AccessibilityIdentifiers.swift
+//  ImageFeed
+//
+//  Created by Денис Кель on 24.03.2025.
+//
+
+struct AccessibilityIdentifiers {
+    struct Auth {
+        static let loginButton = "Authenticate"
+    }
+
+    struct Profile {
+        static let userNameLabel = "userNameLabel"
+        static let loginNameLabel = "loginNameLabel"
+        static let logoutButton = "logoutButton"
+    }
+
+    struct Feed {
+        static let likeButton = "like no active"
+        static let unLikeButton = "like active"
+        static let backButton = "navBackButton"
+    }
+}
