+++ ImageFeed/Auth/WebViewPresenterProtocol.swift
@@ -8,11 +8,12 @@
 import Foundation
 
 
-public protocol WebViewPresenterProtocol {
+protocol WebViewPresenterProtocol {
+    var view: WebViewViewControllerProtocol? { get set }
+    
     func viewDidLoad()
     func didUpdateProgressValue(_ newValue: Double)
     func code(from url: URL) -> String?
-
-    var view: WebViewViewControllerProtocol? { get set }
-
+    
+    
 }
