+++ ImageFeed/Auth/WebViewViewController.swift
@@ -8,23 +8,22 @@
 import UIKit
 @preconcurrency import WebKit
 
-
 final class WebViewViewController: UIViewController & WebViewViewControllerProtocol {
     
     weak var delegate: WebViewViewControllerDelegate?
-    private var webView: WKWebView = WKWebView()
-    private var backButton: UIButton = UIButton()
-    private var progressView: UIProgressView = UIProgressView()
+    private var webView = WKWebView()
+    private var progressView = UIProgressView()
     private var estimatedProgressObservation: NSKeyValueObservation?
-    var presenter: WebViewPresenterProtocol?
     
+    var presenter: WebViewPresenterProtocol?
     
     override func viewDidLoad() {
         super.viewDidLoad()
         setupWebView()
         setupProgressView()
-        webView.navigationDelegate = self
         setupConstraints()
+        
+        webView.navigationDelegate = self
         loadAuthView()
         
         estimatedProgressObservation = webView.observe(
@@ -32,7 +31,7 @@ final class WebViewViewController: UIViewController & WebViewViewControllerProto
              options: [],
              changeHandler: { [weak self] _, _ in
                  guard let self = self else { return }
-                 presenter?.didUpdateProgressValue(webView.estimatedProgress)
+                 self.presenter?.didUpdateProgressValue(webView.estimatedProgress)
              }
         )
     }
@@ -50,20 +49,13 @@ final class WebViewViewController: UIViewController & WebViewViewControllerProto
         view.addSubview(progressView)
     }
     
-    private func loadAuthView() {
-        presenter?.viewDidLoad()
-        
-    }
-    
     private func setupConstraints() {
-        
         NSLayoutConstraint.activate([
-            // Констрейнты для webView
             webView.topAnchor.constraint(equalTo: view.topAnchor),
             webView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
             webView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
             webView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
-            // Констрейнты для progressView
+            
             progressView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
             progressView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor),
             progressView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
@@ -83,20 +75,55 @@ final class WebViewViewController: UIViewController & WebViewViewControllerProto
         progressView.isHidden = isHidden
     }
     
-    deinit {
-        estimatedProgressObservation = nil
-    }
-    
     func load(request: URLRequest) {
         webView.load(request)
     }
     
+    private func loadAuthView() {
+        presenter?.viewDidLoad()
+        
+    }
+    
+    
+    deinit {
+        estimatedProgressObservation = nil
+    }
 }
 
-public protocol WebViewViewControllerProtocol: AnyObject {
+protocol WebViewViewControllerProtocol: AnyObject {
     var presenter: WebViewPresenterProtocol? { get set }
     func load(request: URLRequest)
     func setProgressValue(_ newValue: Float)
     func setProgressHidden(_ isHidden: Bool)
+}
+
+
+extension WebViewViewController: WKNavigationDelegate {
+    func webView(
+        _ webView: WKWebView,
+        decidePolicyFor navigationAction: WKNavigationAction,
+        decisionHandler: @escaping (WKNavigationActionPolicy) -> Void
+    ) {
+        if let code = code(from: navigationAction) {
+            UIBlockingProgressHUD.show()
+            delegate?.webViewViewController(self, didAuthenticateWithCode: code)
+            decisionHandler(.cancel)
+        } else {
+            decisionHandler(.allow)
+        }
+    }
     
+    private func code(from navigationAction: WKNavigationAction) -> String? {
+        if
+            let url = navigationAction.request.url,
+            let urlComponents = URLComponents(string: url.absoluteString),
+            urlComponents.path == "/oauth/authorize/native",
+            let items = urlComponents.queryItems,
+            let codeItem = items.first(where: { $0.name == "code" })
+        {
+            return codeItem.value
+        } else {
+            return nil
+        }
+    }
 }
