+++ ImageFeed/AuthConfiguration.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 enum Constants {
-    static let accessKey = "dakfFfYrzaq7hIb7pltxnalJlaXCReWwCZ3kdqA2QyQ"
-    static let secretKey = "cXLjAdlrNcT1bxkyM5XiiYYEzQQZqCPM0hfxCPNzmXI"
+    static let accessKey = "Onz9XNMFutxChRhZO-jM--5cQV6YjJquT7dewtfeKaw"
+    static let secretKey = "cq3-gQNQZV6rXrlD39RHDkEAApxqe2W4XlujFmdk844"
     static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
     static let accessScope = "public+read_user+write_likes"
     static let defaultBaseURL: URL = {
@@ -30,14 +30,14 @@ struct AuthConfiguration {
     let defaultBaseURL: URL
     let authURLString: String
     static var standard: AuthConfiguration {
-        return AuthConfiguration(accessKey: Constants.accessKey,
-                                 secretKey: Constants.secretKey,
-                                 redirectURI: Constants.redirectURI,
-                                 accessScope: Constants.accessScope,
-                                 authURLString: Constants.unsplashAuthorizeURLString,
-                                 defaultBaseURL: Constants.defaultBaseURL)
+        AuthConfiguration(accessKey: Constants.accessKey,
+                          secretKey: Constants.secretKey,
+                          redirectURI: Constants.redirectURI,
+                          accessScope: Constants.accessScope,
+                          authURLString: Constants.unsplashAuthorizeURLString,
+                          defaultBaseURL: Constants.defaultBaseURL)
     }
-
+    
     init(accessKey: String, secretKey: String, redirectURI: String, accessScope: String, authURLString: String, defaultBaseURL: URL) {
         self.accessKey = accessKey
         self.secretKey = secretKey
