+++ ImageFeed/Base.lproj/Main.storyboard
@@ -50,6 +50,19 @@
                                                     <rect key="frame" x="16" y="4" width="382" height="203"/>
                                                     <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                                 </imageView>
+                                                <button opaque="NO" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="FlR-L1-WkE">
+                                                    <rect key="frame" x="354" y="4" width="44" height="44"/>
+                                                    <accessibility key="accessibilityConfiguration" identifier="LikeButton"/>
+                                                    <constraints>
+                                                        <constraint firstAttribute="width" constant="44" id="3ed-gx-21B"/>
+                                                        <constraint firstAttribute="height" constant="44" id="xcG-q4-wXR"/>
+                                                    </constraints>
+                                                    <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
+                                                    <state key="normal" image="like_active"/>
+                                                    <connections>
+                                                        <action selector="likeButtonClicked" destination="6pT-fg-4yg" eventType="touchUpInside" id="SDB-Za-Rz7"/>
+                                                    </connections>
+                                                </button>
                                                 <view userInteractionEnabled="NO" contentMode="scaleAspectFill" semanticContentAttribute="playback" translatesAutoresizingMaskIntoConstraints="NO" id="kwa-il-mer">
                                                     <rect key="frame" x="16" y="177" width="382" height="30"/>
                                                     <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
@@ -63,18 +76,6 @@
                                                     <color key="textColor" white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                                     <nil key="highlightedColor"/>
                                                 </label>
-                                                <button opaque="NO" contentMode="scaleToFill" insetsLayoutMarginsFromSafeArea="NO" contentHorizontalAlignment="center" contentVerticalAlignment="center" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="FlR-L1-WkE">
-                                                    <rect key="frame" x="354" y="4" width="44" height="44"/>
-                                                    <constraints>
-                                                        <constraint firstAttribute="width" constant="44" id="3ed-gx-21B"/>
-                                                        <constraint firstAttribute="height" constant="44" id="xcG-q4-wXR"/>
-                                                    </constraints>
-                                                    <inset key="imageEdgeInsets" minX="0.0" minY="0.0" maxX="2.2250738585072014e-308" maxY="0.0"/>
-                                                    <state key="normal" image="like_active"/>
-                                                    <connections>
-                                                        <action selector="likeButtonClicked" destination="6pT-fg-4yg" eventType="touchUpInside" id="SDB-Za-Rz7"/>
-                                                    </connections>
-                                                </button>
                                             </subviews>
                                             <color key="backgroundColor" white="0.0" alpha="0.0" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
                                             <constraints>
