+++ ImageFeed/ImagesList/ImagesListViewController+Extensions.swift
@@ -20,8 +20,9 @@ extension ImagesListViewController {
         
         let likeImage = UIImage(named: image.isLiked ? "like_active" : "like_no_active")
         cell.likeButton.setImage(likeImage, for: .normal)
+        cell.likeButton.accessibilityIdentifier = image.isLiked ? AccessibilityIdentifiers.Feed.unLikeButton : AccessibilityIdentifiers.Feed.likeButton
         if let createdAt = image.createdAt {
-            cell.dateLabel.text = dateFormatter.string(from: createdAt)
+            cell.dateLabel.text = ImagesListViewController.dateFormatter.string(from: createdAt)
         } else {
             cell.dateLabel.text = image.welcomeDescription
         }
@@ -89,11 +90,14 @@ extension ImagesListViewController: ImagesListCellDelegate {
                 UIBlockingProgressHUD.dismiss()
                 if success {
                     self.presenter.photos[indexPath.row].isLiked = newLikeStatus
+                    let likeImage = UIImage(named: newLikeStatus ? "like_active" : "like_no_active")
+                    cell.likeButton.setImage(likeImage, for: .normal)
+                    cell.likeButton.accessibilityIdentifier = newLikeStatus ? AccessibilityIdentifiers.Feed.unLikeButton : AccessibilityIdentifiers.Feed.likeButton
                     self.tableView.reloadRows(at: [indexPath], with: .automatic)
                 } else {
                     print("Ошибка при изменении лайка")
                 }
             }
         }
-    }
-}
+    }}
+
