+++ ImageFeedTests/WebViewPresenterSpy.swift
@@ -5,19 +5,28 @@
 //  Created by Денис Кель on 17.03.2025.
 //
 
-import ImageFeed
+@testable import ImageFeed
 import Foundation
 
 final class WebViewPresenterSpy: WebViewPresenterProtocol {
-    var viewDidLoadCalled: Bool = false
-    var view: WebViewViewControllerProtocol?
+    weak var view: WebViewViewControllerProtocol?
+    
+    var viewDidLoadCalled = false
+    var didUpdateProgressValueCalled = false
+    var receivedProgressValue: Double?
+    
     
     func viewDidLoad() {
         viewDidLoadCalled = true
+        let request = URLRequest(url: URL(string: "https://unsplash.com")!)
+        
+        view?.load(request: request)
+        
     }
     
     func didUpdateProgressValue(_ newValue: Double) {
-        
+        didUpdateProgressValueCalled = true
+        receivedProgressValue = newValue
     }
     
     func code(from url: URL) -> String? {
