+++ ImageFeedTests/WebViewTests.swift
@@ -1,0 +1,106 @@
+//
+//  ImageFeedTests.swift
+//  ImageFeedTests
+//
+//  Created by Денис Кель on 17.03.2025.
+//
+
+@testable import ImageFeed
+import XCTest
+import WebKit
+
+final class WebViewTests: XCTestCase {
+    
+    func testViewControllerCallsViewDidLoad() {
+        // Given
+        let viewController = WebViewViewController()
+        let presenter = WebViewPresenterSpy()
+        viewController.presenter = presenter
+        presenter.view = viewController
+        
+        // When
+        viewController.loadViewIfNeeded()
+        
+        // Then
+        XCTAssertTrue(presenter.viewDidLoadCalled)
+    }
+    
+    func testPresenterCallsLoadRequest() {
+        // Given
+        let viewController = WebViewViewControllerSpy()
+        //        let authHelper = AuthHelper()
+        let presenter = WebViewPresenterSpy()
+        viewController.presenter = presenter
+        presenter.view = viewController
+        
+        // When
+        presenter.viewDidLoad()
+        
+        // Then
+        XCTAssertTrue(viewController.loadRequestCalled)
+    }
+    
+    func testProgressVisibleWhenLessThenOne() {
+        // Given
+        let authHelper = AuthHelper()
+        let presenter = WebViewPresenter(authHelper: authHelper)
+        let progress: Float = 0.6
+        
+        // When
+        let shouldHideProgress = presenter.shouldHideProgress(for: progress)
+        
+        // Then
+        XCTAssertFalse(shouldHideProgress)
+    }
+    
+    func testProgressHiddenWhenOne() {
+        // Given
+        let authHelper = AuthHelper()
+        let presenter = WebViewPresenter(authHelper: authHelper)
+        let progress: Float = 1.0
+        
+        // When
+        let shouldHideProgress = presenter.shouldHideProgress(for: progress)
+        
+        // Then
+        XCTAssertTrue(shouldHideProgress)
+    }
+    
+    func testAuthHelperAuthURL() {
+        // Given
+        let configuration = AuthConfiguration.standard
+        let authHelper = AuthHelper(configuration: configuration)
+        
+        // When
+        let url = authHelper.authURL
+        
+        guard let urlString = url?.absoluteString else {
+            XCTFail("Auth URL is nil")
+            return
+        }
+        
+        // Then
+        XCTAssertTrue(urlString.contains(configuration.authURLString))
+        XCTAssertTrue(urlString.contains(configuration.accessKey))
+        XCTAssertTrue(urlString.contains(configuration.redirectURI))
+        XCTAssertTrue(urlString.contains("code"))
+        XCTAssertTrue(urlString.contains(configuration.accessScope))
+    }
+    
+    func testCodeFromURL() {
+        // Given
+        var urlComponents = URLComponents(string: "https://unsplash.com/oauth/authorize/native")!
+        urlComponents.queryItems = [URLQueryItem(name: "code", value: "test code")]
+        guard let url = urlComponents.url else {
+            XCTFail("Invalid URL")
+            return
+        }
+        let authHelper = AuthHelper()
+        
+        // When
+        let code = authHelper.getCode(from: url)
+        
+        // Then
+        XCTAssertEqual(code, "test code")
+    }
+}
