+++ ImageFeedTests/WebViewViewControllerSpy.swift
@@ -6,23 +6,25 @@
 //
 
 import Foundation
-import ImageFeed
+@testable import ImageFeed
 
 
 final class WebViewViewControllerSpy: WebViewViewControllerProtocol {
-    var presenter: ImageFeed.WebViewPresenterProtocol?
-
+    var presenter: WebViewPresenterProtocol?
+    
     var loadRequestCalled: Bool = false
-
+    var setProgressValueCalled: Bool = false
+    var setProgressHiddenCalled: Bool = false
+    
     func load(request: URLRequest) {
         loadRequestCalled = true
     }
-
+    
     func setProgressValue(_ newValue: Float) {
-
+        setProgressValueCalled = true
     }
-
+    
     func setProgressHidden(_ isHidden: Bool) {
-
+        setProgressHiddenCalled = true
     }
 }
