+++ ImageFeedUITests/ImageFeedUITests.swift
@@ -6,99 +6,108 @@
 //
 
 import XCTest
+@testable import ImageFeed
+import ProgressHUD
 
-class ImageFeedUITests: XCTestCase {
+final class ImageFeedUITests: XCTestCase {
     private var app: XCUIApplication!
-
-    override func setUp() {
+    
+    override func setUpWithError() throws {
+        try super.setUpWithError()
         continueAfterFailure = false
-        super.setUp()
         app = XCUIApplication()
         app.launch()
     }
-
-    override func tearDown() {
+    
+    override func tearDownWithError() throws {
+        app.terminate()
         app = nil
-        super.tearDown()
+        try super.tearDownWithError()
     }
-
+    
     func testAuth() throws {
         let authButton = app.buttons["Authenticate"]
         XCTAssertTrue(authButton.waitForExistence(timeout: 5), "Кнопка авторизации не найдена")
         authButton.tap()
-
+        
         let webView = app.webViews["UnsplashWebView"]
         XCTAssertTrue(webView.waitForExistence(timeout: 10), "WebView не загрузился")
-
+        
         let loginTextField = webView.textFields.element
         let passwordTextField = webView.secureTextFields.element
         XCTAssertTrue(loginTextField.waitForExistence(timeout: 10), "Поле логина не найдено")
         XCTAssertTrue(passwordTextField.waitForExistence(timeout: 10), "Поле пароля не найдено")
-
+        
         loginTextField.tap()
-        loginTextField.typeText("--логин--")
+        loginTextField.typeText("asket2013@ya.ru")
         webView.swipeUp()
-
+        
         passwordTextField.tap()
-        passwordTextField.typeText("--пароль--")
-
+        passwordTextField.typeText("Def53917468")
+        
         webView.swipeUp()
         app.buttons["Login"].tap()
-
+        
         let firstCell = app.tables.cells.element(boundBy: 0)
         XCTAssertTrue(firstCell.waitForExistence(timeout: 15), "Лента не загрузилась")
     }
-
+    
+    
     func testFeed() throws {
         let tablesQuery = app.tables
         let firstCell = tablesQuery.cells.element(boundBy: 0)
+        sleep(10)
+        let likeButton = firstCell.buttons[AccessibilityIdentifiers.Feed.likeButton]
+        
+        likeButton.tap()
+        sleep(10)
+        let unLikeButton = firstCell.buttons[AccessibilityIdentifiers.Feed.unLikeButton]
+        unLikeButton.tap()
+        sleep(10)
         firstCell.swipeUp()
-
-        sleep(3)
-        firstCell.buttons["ButtonLike"].tap()
-        sleep(3)
-        firstCell.buttons["ButtonLike"].tap()
-        sleep(3)
-
-        firstCell.swipeUp()
-        sleep(2)
+        sleep(5)
+        firstCell.swipeDown()
+        sleep(5)
+        let cellToImage = tablesQuery.cells.element(boundBy: 0)
+        cellToImage.tap()
+        sleep(10)
+        let image = app.scrollViews.images.element(boundBy: 0)
         
-        firstCell.tap()
-        sleep(4)
-        let imageScrollView = app.scrollViews["ImageScrollView"]
-//        XCTAssertTrue(imageScrollView.waitForExistence(timeout: 5))
-        sleep(3)
-        let image = imageScrollView.images.element(boundBy: 0)
-        sleep(3)
-
-        image.pinch(withScale: 1.25, velocity: 1)
-        sleep(1)
+        image.pinch(withScale: 3, velocity: 1)
+        sleep(10)
         
-        image.pinch(withScale: 0.1, velocity: -1)
-        sleep(1)
+        image.pinch(withScale: 0.5, velocity: -1)
+        sleep(10)
         
-        let backButton = app.buttons["navBackButton"]
+        let backButton = app.buttons[AccessibilityIdentifiers.Feed.backButton]
         backButton.tap()
-
+        
     }
-
+    
     func testProfile() throws {
         let profileTab = app.tabBars.buttons.element(boundBy: 1)
         XCTAssertTrue(profileTab.waitForExistence(timeout: 5))
         profileTab.tap()
-
-        let userNameLabel = app.staticTexts["userNameLabel"]
+        
+        let userNameLabel = app.staticTexts[AccessibilityIdentifiers.Profile.userNameLabel]
         XCTAssertTrue(userNameLabel.waitForExistence(timeout: 5))
-
-        let loginNameLabel = app.staticTexts["loginNameLabel"]
+        
+        let loginNameLabel = app.staticTexts[AccessibilityIdentifiers.Profile.loginNameLabel]
         XCTAssertTrue(loginNameLabel.waitForExistence(timeout: 5))
-
-        let logoutButton = app.buttons["logoutButton"]
+        
+        let logoutButton = app.buttons[AccessibilityIdentifiers.Profile.logoutButton]
         XCTAssertTrue(logoutButton.waitForExistence(timeout: 5))
         logoutButton.tap()
-
+        
         let alert = app.alerts["Пока, пока!"]
         XCTAssertTrue(alert.waitForExistence(timeout: 5))
         alert.scrollViews.otherElements.buttons["Да"].tap()
+    }
+    
+    private func scrollToElementIfNeeded(_ element: XCUIElement, in table: XCUIElement) {
+        while !element.isHittable {
+            table.swipeDown()
+            sleep(2)
+        }
     }
 }
